+++ Events.php
@@ -8,9 +8,12 @@ use humhub\modules\fcmPush\components\NotificationTargetProvider;
 use humhub\modules\fcmPush\helpers\MobileAppHelper;
 use humhub\modules\fcmPush\helpers\WebAppHelper;
 use humhub\modules\fcmPush\services\DriverService;
+use humhub\modules\fcmPush\widgets\RegisterDeviceTokenButton;
 use humhub\modules\notification\targets\MobileTargetProvider;
+use humhub\modules\notification\widgets\NotificationSettingsForm;
 use humhub\modules\web\pwa\controllers\ServiceWorkerController;
 use Yii;
+use yii\base\WidgetEvent;
 
 class Events
 {
@@ -43,12 +46,13 @@ class Events
         $bundle = FirebaseAsset::register(Yii::$app->view);
 
         $pushDriver = (new DriverService($module->getConfigureForm()))->getWebDriver();
+        $baseUrl = Yii::getAlias($bundle->baseUrl);
 
         // Service Worker Addons
         $controller->additionalJs .= <<<JS
             // Give the service worker access to Firebase Messaging.
-            importScripts('{$bundle->baseUrl}/firebase-app-compat.js');
-            importScripts('{$bundle->baseUrl}/firebase-messaging-compat.js');
+            importScripts('{$baseUrl}/firebase-app-compat.js');
+            importScripts('{$baseUrl}/firebase-messaging-compat.js');
 
             firebase.initializeApp({
                 messagingSenderId: "{$pushDriver->getSenderId()}",
@@ -110,5 +114,15 @@ JS;
     {
         Yii::$app->session->set(WebAppHelper::SESSION_VAR_UNREGISTER_NOTIFICATION, 1);
         Yii::$app->session->set(MobileAppHelper::SESSION_VAR_UNREGISTER_NOTIFICATION, 1);
+    }
+
+    public static function onNotificationSettingsFormAfterRun(WidgetEvent $event)
+    {
+        /** @var NotificationSettingsForm $form */
+        $form = $event->sender;
+
+        if ($form->model->user) { // Only show the button for User settings (not admin settings)
+            $event->result = RegisterDeviceTokenButton::widget() . $event->result;
+        }
     }
 }